Fix media import with external urls
Importing media from CSV which contains a hyperlink results in an exception, because the controller does not check correctly if the file exists in the file system.
The force flag on the scale call is used in the media/scale controller to circumvent file system checks, which is required here, too.
The regex is the same as the one in \Aimeos\Controller\Common\Media\Standard::getFileContent used to determine external sources.